+++ middleware/grpc_logging_test.go
@@ -15,12 +15,10 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
-
-	dskit_log "github.com/grafana/dskit/log"
 )
 
 func BenchmarkGRPCServerLog_UnaryServerInterceptor_NoError(b *testing.B) {
-	logger := dskit_log.GoKit(level.NewFilter(log.NewNopLogger(), level.AllowError()))
+	logger := level.NewFilter(log.NewNopLogger(), level.AllowError())
 	l := GRPCServerLog{Log: logger, WithRequest: false, DisableRequestSuccessLog: true}
 	ctx := context.Background()
 	info := &grpc.UnaryServerInfo{FullMethod: "Test"}
@@ -64,7 +62,7 @@ func TestGrpcLogging(t *testing.T) {
 	}} {
 		t.Run("", func(t *testing.T) {
 			buf := bytes.NewBuffer(nil)
-			logger := dskit_log.GoKit(log.NewLogfmtLogger(buf))
+			logger := log.NewLogfmtLogger(buf)
 			l := GRPCServerLog{Log: logger, WithRequest: true, DisableRequestSuccessLog: false}
 
 			handler := func(ctx context.Context, req interface{}) (interface{}, error) {